refactor(metadata): split test module into per-domain *_tests.rs files
Replace the single 2180-line #[cfg(test)] mod tests block in mod.rs with
12 per-domain sibling test files sharing fixtures via a new test_support.rs:

  test_support.rs               shared fixtures (sample_profile, connection,
                                clean_exit_report, insert_test_profile_row)
  profile_sync_tests.rs         profile write/rename/delete/sync + symlink
  launcher_tests.rs             launcher + launch_history (exported/renamed/
                                deleted, launch started/finished, sweep,
                                diagnostic JSON truncation, disabled noop)
  community_index_tests.rs      tap indexing, watermark skip, FK cascade
                                (keeps local sample_tap_workspace/entry/
                                sync_result helpers)
  cache_tests.rs                cache entries (round-trip, idempotent,
                                oversized, expiry, disabled noop)
  collections_crud_tests.rs     CRUD + rename + description + cascade
  collections_defaults_tests.rs per-collection launch defaults roundtrip
  collections_favorites_tests.rs  favorite toggles + excludes-deleted
  launch_queries_tests.rs       usage insights (most_launched, failure
                                trends, single-profile queries)
  migrations_sanity_tests.rs    v9→v10 GPU presets, v8→v9 version_snapshots
  trainer_hash_tests.rs         verify cache, launch-check baseline/mismatch,
                                launch-issues mapping
  version_store_tests.rs        upsert/lookup/pruning/acknowledge/bulk-load
                                + disabled noop
  correlation_status_tests.rs   compute_correlation_status pure fn tests

Test count identical before/after: 1078 passed, 0 failed. Zero code move in
test bodies — each #[test] fn is byte-identical. Sample helpers scoped to
test_support.rs are pub(super); community helpers stay local to
community_index_tests.rs since they're used nowhere else.

metadata/mod.rs: 2238 → 91 lines (-2147). Every new test file is ≤ 500 lines
per the CLAUDE.md rule. Collections tests split into 3 files (crud/defaults/
favorites) to keep them under the limit.

Phases 4-5 of 6 for the metadata/mod.rs split.

Refs #291
Refs #290
